+++ src/modules/types/questionTypes.ts
@@ -1,0 +1,8 @@
+import Inference from "../../inference";
+import ProfitLoss from "../../profitLoss";
+import Tsurukame from "../../tsurukame";
+import WorkCalc from "../../workCalc";
+
+type questionTypes = Tsurukame | Inference | ProfitLoss | WorkCalc;
+
+export default questionTypes;
